17052023_1610 Crm - Print & Jobs

+++ GaCloudServer.Resources.Api/Controllers/GaCrmController.cs
@@ -1,12 +1,18 @@
 ﻿using AutoWrapper.Wrappers;
 using GaCloudServer.BusinnessLogic.Dtos.Resources.Crm;
+using GaCloudServer.BusinnessLogic.Dtos.Template;
+using GaCloudServer.BusinnessLogic.Services;
 using GaCloudServer.BusinnessLogic.Services.Interfaces;
 using GaCloudServer.Resources.Api.Configuration.Constants;
 using GaCloudServer.Resources.Api.Dtos.Crm;
+using GaCloudServer.Resources.Api.Dtos.Custom;
+using GaCloudServer.Resources.Api.Dtos.Resources.ContactCenter;
 using GaCloudServer.Resources.Api.ExceptionHandling;
 using GaCloudServer.Resources.Api.Mappers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Graph;
+using code = Microsoft.AspNetCore.Http.StatusCodes;
 
 namespace GaCloudServer.Resources.Api.Controllers
 {
@@ -19,13 +25,16 @@ namespace GaCloudServer.Resources.Api.Controllers
     {
         private readonly IGaCrmService _gaCrmService;
         private readonly ILogger<GaCrmController> _logger;
+        private readonly IPrintService _printService;
 
         public GaCrmController(
             IGaCrmService gaCrmService
+            , IPrintService printService
             , ILogger<GaCrmController> logger)
         {
 
             _gaCrmService = gaCrmService;
+            _printService = printService;
             _logger = logger;
         }
 
@@ -344,6 +353,23 @@ namespace GaCloudServer.Resources.Api.Controllers
 
         }
 
+        [HttpGet("GetGaCrmEventsByBoardAsync/{date}/{area}")]
+        public async Task<ActionResult<ApiResponse>> GetGaCrmEventsByBoardAsync(DateTime date,long area)
+        {
+            try
+            {
+                var dtos = await _gaCrmService.GetGaCrmEventByBoardAsync(date,area);
+                var apiDtos = dtos.ToApiDto<CrmEventsApiDto, CrmEventsDto>();
+                return new ApiResponse(apiDtos);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message, ex);
+                throw new ApiException(ex.Message);
+            }
+
+        }
+
         [HttpGet("GetGaCrmEventByIdAsync/{id}")]
         public async Task<ActionResult<ApiResponse>> GetGaCrmEventByIdAsync(long id)
         {
@@ -377,6 +403,7 @@ namespace GaCloudServer.Resources.Api.Controllers
             }
 
         }
+
 
         [HttpPost("AddGaCrmEventAsync")]
         public async Task<ActionResult<ApiResponse>> AddGaCrmEventAsync([FromBody] CrmEventApiDto apiDto)
@@ -432,9 +459,19 @@ namespace GaCloudServer.Resources.Api.Controllers
         {
             try
             {
-                var response = await _gaCrmService.DeleteGaCrmEventAsync(id);
+                var devicesRemove = await _gaCrmService.DeleteGaCrmEventDevicesByEventIdAsync(id);
+                if (devicesRemove)
+                {
+                    var response = await _gaCrmService.DeleteGaCrmEventAsync(id);
 
-                return new ApiResponse(response);
+                    return new ApiResponse(response);
+                }
+                else
+                {
+                    return new ApiResponse(false);
+                }
+
+                
             }
             catch (Exception ex)
             {
@@ -445,42 +482,76 @@ namespace GaCloudServer.Resources.Api.Controllers
         }
 
         #region Functions
-        //[HttpGet("ValidateGaCrmEventAsync/{id}/{descrizione}")]
-        //public async Task<ActionResult<ApiResponse>> ValidateGaCrmEventAsync(long id, string descrizione)
-        //{
-        //    try
-        //    {
-        //        var response = await _gaCrmService.ValidateGaCrmEventAsync(id, descrizione);
-        //        return new ApiResponse(response);
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        _logger.LogError(ex.Message, ex);
-        //        throw new ApiException(ex.Message);
-        //    }
+        
 
-        //}
+        [HttpGet("PrintGaCrmEventsByFilterAsync/{date}/{areaId}")]
+        public async Task<ApiResponse> PrintGaContactCenterIngByFilterAsync(DateTime date,long areaId)
+        {
+            try
+            {
+                var list = await _gaCrmService.GetGaCrmEventByBoardAsync(date,areaId);
 
-        //[HttpGet("ChangeStatusGaCrmEventAsync/{id}")]
-        //public async Task<ActionResult<ApiResponse>> ChangeStatusGaCrmEventAsync(long id)
-        //{
-        //    try
-        //    {
-        //        var response = await _gaCrmService.ChangeStatusGaCrmEventAsync(id);
-        //        return new ApiResponse(response);
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        _logger.LogError(ex.Message, ex);
-        //        throw new ApiException(ex.Message);
-        //    }
+                if (list == null || (list.Data.Where(x => x.CrmEventStateId == 1).Count() == 0))
+                {
+                    return new ApiResponse("NoData", null, code.Status200OK);
+                }
 
-        //}
+                var area = await _gaCrmService.GetGaCrmEventAreaByIdAsync(areaId);
+                var dto = await  GenerateCrmEventsTemplate(list.Data, date, area.Descrizione);
+                var response = await _printService.Print("CrmEvents", dto);
+
+                //var dto = GenerateContactCenterTicketsIngTemplate(apiDto.all ? list : list.Where(x => x.Stato == "IN GESTIONE").ToList(), apiDto.comuneAltro, apiDto.dataEsecuzione.GetValueOrDefault(), apiDto.tipoStampa);
+                //var response = await _printService.Print("ContactCenterTicketsIng", dto);
+
+
+                //var printDto = new ContactCenterStatusTicketsApiDto();
+                //if (apiDto.all)
+                //{
+                //    printDto.ticketsId = (from x in list
+                //                          select x.Id).ToArray();
+
+                //}
+                //else
+                //{
+                //    printDto.ticketsId = (from x in list
+                //                          where x.Stato == "IN GESTIONE"
+                //    select x.Id).ToArray();
+
+                //}
+                //await _gaContactCenterService.SetPrintedGaContactCenterTicketsAsync(printDto.ticketsId);
+
+                return new ApiResponse(response);
+
+            }
+            catch (Exception ex)
+            {
+                throw new ApiProblemDetailsException(code.Status400BadRequest);
+            }
+        }
         #endregion
 
         #endregion
 
         #region CrmEventDevices
+        [HttpGet("GetGaCrmEventDevicesByEventIdAsync/{id}")]
+        public async Task<ActionResult<ApiResponse>> GetGaCrmEventDevicesByEventIdAsync(long id)
+        {
+            try
+            {
+                var dto = await _gaCrmService.GetGaCrmEventDevicesByEventIdAsync(id);
+                var apiDto = dto.ToApiDto<CrmEventDevicesApiDto, CrmEventDevicesDto>();
+                return new ApiResponse(apiDto);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message, ex);
+                throw new ApiException(ex.Message);
+            }
+
+        }
+
+
+        #region Functions
         [HttpGet("ChangeStatusGaCrmEventDeviceAsync/{id}")]
         public async Task<ActionResult<ApiResponse>> ChangeStatusGaCrmEventDeviceAsync(long id)
         {
@@ -512,6 +583,75 @@ namespace GaCloudServer.Resources.Api.Controllers
                 _logger.LogError(ex.Message, ex);
                 throw new ApiException(ex.Message);
             }
+
+        }
+        #endregion
+        #endregion
+
+        #region CrmEventJobs
+        [HttpGet("GetViewGaCrmEventJobsAsync")]
+        public async Task<ActionResult<ApiResponse>> GetViewGaCrmEventJobsAsync()
+        {
+            try
+            {
+                var view = await _gaCrmService.GetViewGaCrmEventJobsAsync();
+                return new ApiResponse(view);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message, ex);
+                throw new ApiException(ex.Message);
+            }
+
+        }
+
+        [HttpPost("GetViewGaCrmEventJobsByFilterAsync")]
+        public async Task<ActionResult<ApiResponse>> GetViewGaCrmEventJobsByFilterAsync([FromBody]DateRangeDto filter)
+        {
+            try
+            {
+                var view = await _gaCrmService.GetViewGaCrmEventJobsByFilterAsync(filter.dateStart,filter.dateEnd);
+                return new ApiResponse(view);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message, ex);
+                throw new ApiException(ex.Message);
+            }
+
+        }
+
+        #endregion
+
+        #region Helpers
+        private async Task<CrmEventsTemplateDto> GenerateCrmEventsTemplate(List<CrmEventDto> events, DateTime date, string area)
+        {
+            var dto = new CrmEventsTemplateDto()
+            {
+                FileName = "CrmEventss.pdf",
+                FilePath = @"Print/Crm",
+                Title = "Crm Eventi Programmati",
+                Css = "CrmEvents",
+                Area = area,
+                Data = date.ToString("dd/MM/yyyy"),
+                Items = new List<CrmEventDto>(),
+                Devices= new List<CrmEventDeviceDto>()
+            };
+
+            foreach (var item in events)
+            {
+                dto.Items.Add(item);
+
+                var devices = await _gaCrmService.GetGaCrmEventDevicesByEventIdAsync(item.Id);
+                foreach (var device in devices.Data)
+                {
+                    if (device.Selected)
+                        dto.Devices.Add(device);
+                }
+            
+            }
+
+            return dto;
 
         }
         #endregion
